+++ Editor/GUI/AddressableAssetsSettingsGroupEditor.cs
@@ -29,7 +29,21 @@ namespace UnityEditor.AddressableAssets.GUI
 
         SearchField m_SearchField;
         const int k_SearchHeight = 20;
-        internal AddressableAssetSettings settings { get { return AddressableAssetSettingsDefaultObject.Settings; } }
+
+        AddressableAssetSettings m_Settings;
+        internal AddressableAssetSettings settings
+        {
+            get
+            {
+                if (m_Settings == null)
+                {
+                    m_Settings = AddressableAssetSettingsDefaultObject.Settings;
+                }
+
+                return m_Settings;
+            }
+            set => m_Settings = value;
+        }
 
         bool m_ResizingVerticalSplitter;
         Rect m_VerticalSplitterRect = new Rect(0, 0, 10, k_SplitterWidth);
@@ -226,15 +240,7 @@ namespace UnityEditor.AddressableAssets.GUI
                     var searchString = m_SearchField.OnGUI(searchRect, baseSearch, m_SearchStyles[0], m_SearchStyles[1], m_SearchStyles[2]);
                     if (baseSearch != searchString)
                     {
-                        if (ProjectConfigData.hierarchicalSearch)
-                        {
-                            m_EntryTree.customSearchString = searchString;
-                            Reload();
-                        }
-                        else
-                        {
-                            m_EntryTree.searchString = searchString;
-                        }
+                        m_EntryTree?.Search(searchString);
                     }
                 }
             }
@@ -367,8 +373,8 @@ namespace UnityEditor.AddressableAssets.GUI
         {
             if (settings == null)
                 return false;
-				
-			if (!m_ModificationRegistered)
+
+            if (!m_ModificationRegistered)
             {
                 m_ModificationRegistered = true;
                 settings.OnModification -= OnSettingsModification; //just in case...