+++ Editor/Settings/AddressableAssetGroupSchema.cs
@@ -12,9 +12,7 @@ namespace UnityEditor.AddressableAssets.Settings
     /// </summary>
     public class AddressableAssetGroupSchema : ScriptableObject
     {
-        [FormerlySerializedAs("m_group")]
-        [HideInInspector]
-        [SerializeField]
+        [FormerlySerializedAs("m_group")] [HideInInspector] [SerializeField]
         AddressableAssetGroup m_Group;
 
         /// <summary>
@@ -22,10 +20,7 @@ namespace UnityEditor.AddressableAssets.Settings
         /// </summary>
         public AddressableAssetGroup Group
         {
-            get
-            {
-                return m_Group;
-            }
+            get { return m_Group; }
             internal set
             {
                 m_Group = value;
@@ -81,6 +76,7 @@ namespace UnityEditor.AddressableAssets.Settings
                     EditorUtility.SetDirty(this);
                     AddressableAssetUtility.OpenAssetIfUsingVCIntegration(this);
                 }
+
                 if (m_Group != null)
                     m_Group.SetDirty(AddressableAssetSettings.ModificationEvent.GroupSchemaModified, this, postEvent, false);
             }
@@ -105,7 +101,7 @@ namespace UnityEditor.AddressableAssets.Settings
             foreach (var schema in otherSchemas)
             {
                 var s_prop = (new SerializedObject(schema)).FindProperty(propertyName);
-                if ((type == typeof(Enum) && (property.enumValueIndex != s_prop.enumValueIndex)) ||
+                if ((type.IsEnum && (property.enumValueIndex != s_prop.enumValueIndex)) ||
                     (type == typeof(string) && (property.stringValue != s_prop.stringValue)) ||
                     (type == typeof(int) && (property.intValue != s_prop.intValue)) ||
                     (type == typeof(bool) && (property.boolValue != s_prop.boolValue)))
@@ -116,7 +112,9 @@ namespace UnityEditor.AddressableAssets.Settings
 
                 if (type == typeof(ProfileValueReference))
                 {
-                    var field = property.serializedObject.targetObject.GetType().GetField(property.name, BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.DeclaredOnly);
+                    var field = property.serializedObject.targetObject.GetType().GetField(property.name,
+                        BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance |
+                        BindingFlags.DeclaredOnly);
 
                     string lhsId = (field?.GetValue(property.serializedObject.targetObject) as ProfileValueReference)?.Id;
                     string rhsId = (field?.GetValue(s_prop.serializedObject.targetObject) as ProfileValueReference)?.Id;
@@ -130,7 +128,9 @@ namespace UnityEditor.AddressableAssets.Settings
 
                 if (type == typeof(SerializedType))
                 {
-                    var field = property.serializedObject.targetObject.GetType().GetField(property.name, BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance | BindingFlags.DeclaredOnly);
+                    var field = property.serializedObject.targetObject.GetType().GetField(property.name,
+                        BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance |
+                        BindingFlags.DeclaredOnly);
 
                     Type lhs = ((SerializedType)field?.GetValue(property.serializedObject.targetObject)).Value;
                     Type rhs = ((SerializedType)field?.GetValue(s_prop.serializedObject.targetObject)).Value;